+++ src/model/ConfigPreset.ts
@@ -168,6 +168,8 @@ export interface ConfigPreset {
     nemesisSeedFolder?: string; // Optional seed folder if user provides an external seed/00000 folder.
     assemblies?: string;
     databaseName: string;
+    nonVotingUnfinalizedBlocksDuration: string;
+    votingUnfinalizedBlocksDuration?: string;
     nemesisSignerPublicKey: string;
     nemesisGenerationHashSeed: string;
     harvestNetworkFeeSinkAddress?: string;